Fixed play rating auto-rounding and Recent 10 average auto-rounding

+++ VanillaForKonata/BotFunction/Games/Arcaea/Models/Best30.cs
@@ -349,7 +349,7 @@ namespace VanillaForKonata.BotFunction.Games.Arcaea.Models
             /// <summary>
             /// 
             /// </summary>
-            public long recent10_avg { get; set; }
+            public double recent10_avg { get; set; }
             /// <summary>
             /// 
             /// </summary>
